+++ backend/src/context.ts
@@ -1,0 +1,9 @@
+import { PrismaClient } from "@prisma/client";
+
+const prisma = new PrismaClient();
+
+export interface Context {
+  prisma: PrismaClient;
+}
+
+export const context: Context = { prisma: prisma };
